refactor(app-workflows): move files
