--- packages/app-workflows/src/Components/Workflow/index.tsx
@@ -1,5 +1,0 @@
-export * from "./Workflows.js";
-export * from "./WorkflowApplications.js";
-export * from "./WorkflowEditor.js";
-export * from "./Workflow.js";
-export * from "./WorkflowView.js";
