+++ packages/app-workflows/src/Components/Workflows/WorkflowEditorView.tsx
@@ -1,15 +1,15 @@
 import React, { useCallback } from "react";
-import { Workflow } from "./Workflow.js";
+import { WorkflowEditorSteps } from "./WorkflowEditorSteps.js";
 import type { IWorkflowsPresenter } from "~/Presenters/index.js";
 import { Button, Grid, Heading, Loader } from "@webiny/admin-ui";
 import { observer } from "mobx-react-lite";
 import { WorkflowError } from "./Error/WorkflowError.js";
 
-interface WorkflowViewProps {
+interface WorkflowEditorViewProps {
     presenter: IWorkflowsPresenter;
 }
 
-export const WorkflowView = observer((props: WorkflowViewProps) => {
+export const WorkflowEditorView = observer((props: WorkflowEditorViewProps) => {
     const { presenter } = props;
 
     const saveWorkflow = useCallback(() => {
@@ -41,7 +41,7 @@ export const WorkflowView = observer((props: WorkflowViewProps) => {
             <Heading level={2}>{presenter.vm.app.name}</Heading>
             <WorkflowError error={presenter.vm.error} />
             <Grid.Column span={12}>
-                <Workflow presenter={presenter} />
+                <WorkflowEditorSteps presenter={presenter} />
             </Grid.Column>
             <Grid.Column span={12} className={"wby-text-right"}>
                 <Button
