+++ packages/app-workflows/src/hooks/canUseWorkflows.ts
@@ -1,0 +1,11 @@
+import { useWcp } from "@webiny/app-admin";
+
+export const useCanUseWorkflows = () => {
+    const wcp = useWcp();
+
+    const canUseWorkflows = wcp.canUseWorkflows();
+
+    return {
+        canUseWorkflows
+    };
+};
